Reorganize code into smaller files
