+++ secrets/pyproject.toml
@@ -30,7 +30,7 @@ dependencies = [
   "ruamel.yaml~=0.18",
 ]
 [project.scripts]
-spack-secrets = "spack_secrets.cli:main"
+spack-secrets = "spack_secrets.cli:cli"
 
 [tool.hatch.build]
 packages = ["src/spack_secrets"]
